rename: 파일 네이밍 수정
- 컴포넌트 관련 파일 index.js와 style.js로 수정
- 파일 네이밍 수정에 따른 경로 수정

related to #108
