+++ fe/src/components/common/layout/PageLayout.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 
+// 컨테이너 수정
 const PageLayoutContainer = styled.main`
   display: flex;
   flex-direction: column;
