+++ fe/src/components/issueList/issueTable/list/listItems/listItemLeft/ListItemLeft.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import Checkbox from '../../../../../../components/common/checkbox/Checkbox';
+import Checkbox from '@components/common/checkbox';
 import ListItemLeftContent from './listItemLeftContent/ListItemLeftContent';
 
 const ListItemLeftContainer = styled.div`
